// types/CompanyProfile.ts
export interface CompanyProfile {
    name: string;
    companyWebsite: string;

    description?: string;
    industry?: string;
    competitors?: string[];
    targetAudience?: string;
    products?: string;
    benefits?: string;
    values?: string;
    regulatoryContext?: string;
    personas?: string;
    geoSpecificity?: string;

}